Start project 'fancyhotel'
This commit start project 'fancyhotel', so other contributors have
a skeleton to work with. It is still connected to SQLite as its
database, but that's simply because I could not get NTNU's
MySQL-servers to connect as they should. I'll check that out with
'Orakeltjenesten' on Wednesday, I think.

We must later add a settings_secrets-file for database information
and a secret key, but they will be kept here in the open as of now
- mainly because I'm leaving for a cabin trip tomorrow morning and
won't be back until Tuesday, so I don't want to find a way to keep
settings_secrets.py up-to-date with everyone while staying in a
cabin without power and WiFi!

Related to #2 - but I don't consider the issue to be done and ready
for closing until MySQL is used for the database. Luckily, I do think
Django makes changing databases pretty easy - so it shouldn't be
tooooooooo bad!