Change database to MySQL
This commit changes the database tech from SQLite to MySQL, running
on a Google server. Some login-settings should probably be put in
settings_secrets.py, but since the GitLabs-project is private (and
to allow "veiledere" to keep up to date with the project), we
decided to keep the settings in the open :)

This closes #2, as we can add elements to the database without
committing. We are now ready to get going!