Change handling of number of beds in booking.models.Hotelroom
This commit changes handling of number of beds in class Hotelroom.
The Admin now has to add number of beds (in total) and number of
single beds, meaning those are stored in the database and filterable,
while number of double beds are auto-calculated (so they can be fetched
with room.doubleBeds, but not filterable.

Related to general work, possibly to #2, but that's closed!

I think this is a good time to start using branches :)