Add ability to search for available Hotelrooms at /booking/room/
This commit adds the ability to search for available Hotelrooms,
and will display a (rather ugly) list of free room numbers. Please
note that the front-end parts are not in any way done yet!

Also adds room/<roomNr>, which is a site for each room. Little to
no information there as of yet, but it might be a "staging ground"
for bookings, potentially. Still cool to have, I guess - maybe even
with a description field for each room?

Related to #14