Add way of registering bookings with forms in roombooking.html
Note: I have to go, but I will finish this tonight!

This commit enables the user to register a new booking (using the
form BookingForm), with all necessary fields. It is currently VERY
ugly, and doesn't actually register anything (it redirects to room
202 when form is valid, and I haven't actually checked if that works
as intended), but I'm running out of time and have to go!