Change button in URL booking/room/roomNr to go to booking-view
This commit changes the button in booking/room/roomNr (which is
gotten from booking/room), to redirect to the booking-view. The
next step will be to make it auto-fill roomNr (and date), so the user
only has to enter firstName, lastName, email and phoneNr (which will
be auto-filled if the user is logged in).

Related to #7