Add autofill to BookingForm in view roombooking
This commit adds autofill to BookingForm in view roombooking. This
means that the only thing the user has to fill out is their phone
number, as I do not want to override User class at this stage.

Please note that this is related to #5. It does not yet close it, as
the User class does not store phone numbers, but it is veeeeeery close!

It also does not ask the customer to register after booking a room
(where it should auto-fill information into register form), but
that should be doable (just not now, going to bed)