Add autofill of user information in user signup
This commit adds autofill to user information. If there has been
made a booking from a someone not logged in, in this session, then
the values for email, first and last name are stored in the session.
These are then auto-filled in the user registration form, if they
go there.

The thought here is that this makes it so we can simply add a "book
and register"-button to roombooking, which does saves a booking (as
the button does now) and redirects the user to the register page, where
the information (as of this commit) is autofilled.

Related to #5