Add booking overview to Admin
This commit adds a booking overview to Admin. The overview is a table,
with rooms going down and booking status going right. It has hover
text, showing who has booked the room, and is clickable, to easily
take you to edit the correct booking.

Related to #12, not completely done yet