Add highlight of weekends, and some other HTML+CSS-cleanup
This commit adds highlighting of weekends to booking overview. It
simply adds a darker tint (a dark background with opacity = 0.05)
to the columns representing Saturday and Sunday.

This commit also fixes some mistakes in the overview-view. Most
importantly, it makes it so that bookings are started the day the
guest shows up, then has the correct length. This means that the day
the guest checks out is not marked. This makes sense, as you could
potentially get a new guest in that day, even for just one day. It
also makes coding easier, as we have allowed bookings to have the
same check-in-date as someone checks out.

Fixes #12 (at least mostly, might be some minor cleanups left to do,
but they're mostly if-anyone-feels-like-it tasks)