Add Advanced Options to booking.views.getRooms
This commit adds avanced options to booking.views.getRooms, which
means that a customer can now filter rooms based on every filter-able
field the Hotelroom-model has. This means the customer can filter
rooms based on:
 - date, number of beds, max price (as before)
 - single beds
 - breakfast included
 - parking included
 - cancelling included
 - smoking allowed

Please note: In order to make the checkboxes appear checked if the
default value is True, the booking/search.html is pretty hardcoded,
as are both the form and view. This means that any changes to one
of them means that the rest MUST be updated! This is pretty important!

Fixes #14 :)